requirements.txt
setuptools>=18.5
python-telegram-bot>=8.0.0
requests==2.12.4
six==1.10.0
python-dotenv==0.9.1
